+++ CHANGES.md
@@ -1,5 +1,5 @@
 # 0.2
-- Remove support for Python 2
+- Drop support for Python 2
 
 # 0.1
 - Prepare for public release