+++ CHANGES.md
@@ -1,3 +1,6 @@
+# 0.3
+- Support nested subparsers
+
 # 0.2
 - Drop support for Python 2
 - Show available commands